+++ services/worker/src/worker/job_runners/split/first_rows_from_streaming.py
@@ -23,7 +23,6 @@ from libcommon.processing_graph import ProcessingStep
 from libcommon.queue import JobInfo
 from libcommon.simple_cache import DoesNotExist, SplitFullName, get_response
 from libcommon.storage import StrPath
-from libcommon.utils import orjson_dumps
 
 from worker.config import AppConfig, FirstRowsConfig
 from worker.features import get_cell_value
@@ -35,8 +34,8 @@ from worker.job_runner import (
 )
 from worker.job_runners._datasets_based_job_runner import DatasetsBasedJobRunner
 from worker.utils import (
-    RowItem,
     SplitFirstRowsResponse,
+    create_truncated_row_items,
     get_json_size,
     to_features_list,
 )
@@ -210,118 +209,6 @@ def get_rows(
     else:
         logging.debug(f"the rows in the split have been truncated ({rows_max_number} rows)")
     return rows_plus_one[:rows_max_number]
-
-
-# from https://stackoverflow.com/a/43848928/7351594
-def utf8_lead_byte(b: int) -> bool:
-    """A UTF-8 intermediate byte starts with the bits 10xxxxxx."""
-    return (b & 0xC0) != 0x80
-
-
-def utf8_byte_truncate(text: str, max_bytes: int) -> str:
-    """If text[max_bytes] is not a lead byte, back up until a lead byte is
-    found and truncate before that character."""
-    utf8 = text.encode("utf8")
-    if len(utf8) <= max_bytes:
-        return text
-    i = max_bytes
-    while i > 0 and not utf8_lead_byte(utf8[i]):
-        i -= 1
-    return utf8[:i].decode("utf8", "ignore")
-
-
-# Mutates row_item, and returns it anyway
-def truncate_row_item(row_item: RowItem, min_cell_bytes: int) -> RowItem:
-    row = {}
-    for column_name, cell in row_item["row"].items():
-        # for now: all the cells above min_cell_bytes are truncated to min_cell_bytes
-        # it's done by replacing the cell (which can have any type) by a string with
-        # its JSON serialization, and then truncating it to min_cell_bytes
-        cell_json = orjson_dumps(cell)
-        if len(cell_json) <= min_cell_bytes:
-            row[column_name] = cell
-        else:
-            cell_json_str = cell_json.decode("utf8", "ignore")
-            row_item["truncated_cells"].append(column_name)
-            row[column_name] = utf8_byte_truncate(text=cell_json_str, max_bytes=min_cell_bytes)
-    row_item["row"] = row
-    return row_item
-
-
-COMMA_SIZE = 1  # the comma "," is encoded with one byte in utf-8
-
-
-# Mutates row_items, and returns them anyway
-def truncate_row_items(row_items: List[RowItem], min_cell_bytes: int, rows_max_bytes: int) -> List[RowItem]:
-    # compute the current size
-    rows_bytes = sum(get_json_size(row_item) for row_item in row_items) + COMMA_SIZE * (len(row_items) - 1)
-
-    # Loop backwards, so that the last rows are truncated first
-    for row_item in reversed(row_items):
-        if rows_bytes < rows_max_bytes:
-            break
-        previous_size = get_json_size(row_item) + COMMA_SIZE
-        row_item = truncate_row_item(row_item=row_item, min_cell_bytes=min_cell_bytes)
-        new_size = get_json_size(row_item) + COMMA_SIZE
-        rows_bytes += new_size - previous_size
-        row_idx = row_item["row_idx"]
-        logging.debug(f"the size of the rows is now ({rows_bytes}) after truncating row idx={row_idx}")
-    return row_items
-
-
-def to_row_item(row_idx: int, row: Row) -> RowItem:
-    return {
-        "row_idx": row_idx,
-        "row": row,
-        "truncated_cells": [],
-    }
-
-
-def create_truncated_row_items(
-    rows: List[Row],
-    min_cell_bytes: int,
-    rows_max_bytes: int,
-    rows_min_number: int,
-) -> List[RowItem]:
-    row_items = []
-    rows_bytes = 0
-
-    # two restrictions must be enforced:
-    # - at least rows_min_number rows
-    # - at most rows_max_bytes bytes. Note that it's the limit to the sum of the rows sizes. The JSON response size
-    #   will be greater, due to the other fields (row_idx, truncated_cells, features, etc.).
-    # To enforce this:
-    # 1. first get the first rows_min_number rows
-    for row_idx, row in enumerate(rows[:rows_min_number]):
-        row_item = to_row_item(row_idx=row_idx, row=row)
-        rows_bytes += get_json_size(row_item) + COMMA_SIZE
-        row_items.append(row_item)
-
-    # 2. if the total is over the bytes limit, truncate the values, iterating backwards starting
-    # from the last rows, until getting under the threshold
-    # caveat: the truncation might not be enough to get under the threshold if:
-    # - the number of columns is too high
-    # - rows_max_bytes is too low (or even negative)
-    if rows_bytes >= rows_max_bytes:
-        logging.debug(
-            f"the size of the first {rows_min_number} rows ({rows_bytes}) is above the max number of bytes"
-            f" ({rows_max_bytes}), they will be truncated"
-        )
-        return truncate_row_items(row_items=row_items, min_cell_bytes=min_cell_bytes, rows_max_bytes=rows_max_bytes)
-
-    # 3. else: add the remaining rows until the end, or until the bytes threshold
-    for idx, row in enumerate(rows[rows_min_number:]):
-        row_idx = rows_min_number + idx
-        row_item = to_row_item(row_idx=row_idx, row=row)
-        rows_bytes += get_json_size(row_item) + COMMA_SIZE
-        if rows_bytes >= rows_max_bytes:
-            logging.debug(
-                f"the rows in the split have been truncated to {row_idx} row(s) to keep the size"
-                f" ({rows_bytes}) under the limit ({rows_max_bytes})"
-            )
-            break
-        row_items.append(row_item)
-    return row_items
 
 
 def transform_rows(
